feat: refactor and add basic form validation
